tools/autoupdater-upgrader/requirements.txt
rich
PyGithub

